Extract pictures logic in picturesService.js
