+++ src/Presentation/SpaWeb/src/services/picturesService.js
@@ -1,0 +1,13 @@
+const createImageObj = (pictures) => {
+  let data = [];
+  pictures.forEach((element) => {
+    data.push({
+      original: element.url,
+      thumbnail: element.url,
+    });
+  });
+
+  return data;
+};
+
+export default createImageObj;
